% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver_gradient.R
\name{.backtracking_line_search}
\alias{.backtracking_line_search}
\title{Backtracking line search}
\usage{
.backtracking_line_search(
  loglike,
  theta,
  direction,
  max_step,
  backtrack_ratio,
  min_step,
  constraint
)
}
\description{
Backtracking line search
}
\keyword{internal}
