% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_optimize.R
\name{.make_convergence_checker}
\alias{.make_convergence_checker}
\title{Create convergence checker function}
\usage{
.make_convergence_checker(config)
}
\arguments{
\item{config}{Configuration object}
}
\value{
Function that checks convergence
}
\description{
Create convergence checker function
}
\keyword{internal}
