% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_optimize.R
\name{.mle_optimize_direction}
\alias{.mle_optimize_direction}
\title{Internal direction-based optimizer}
\usage{
.mle_optimize_direction(
  loglike,
  direction_fn,
  theta0,
  config,
  constraint,
  use_linesearch
)
}
\arguments{
\item{loglike}{Log-likelihood function}

\item{direction_fn}{Function computing search direction}

\item{theta0}{Initial parameters}

\item{config}{Configuration object (mle_config or subclass)}

\item{constraint}{Domain constraints (mle_constraint object)}

\item{use_linesearch}{Whether to use backtracking line search}
}
\value{
mle_numerical object with optimization results
}
\description{
Core optimization algorithm used by gradient-based solvers.
This is an internal function not meant to be called directly by users.
}
\keyword{internal}
