% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_optimize.R
\name{.print_iteration}
\alias{.print_iteration}
\title{Print iteration information}
\usage{
.print_iteration(iter, theta, direction, loglike)
}
\arguments{
\item{iter}{Iteration number}

\item{theta}{Current parameters}

\item{direction}{Search direction}

\item{loglike}{Log-likelihood function (can be NULL)}
}
\description{
Print iteration information
}
\keyword{internal}
