% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{finalize_trace}
\alias{finalize_trace}
\title{Finalize trace recorder into trace data}
\usage{
finalize_trace(recorder)
}
\arguments{
\item{recorder}{Trace recorder}
}
\value{
List of trace data or NULL
}
\description{
Finalize trace recorder into trace data
}
\keyword{internal}
