% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem.R
\name{get_fisher}
\alias{get_fisher}
\title{Get Fisher information function from problem}
\usage{
get_fisher(problem)
}
\arguments{
\item{problem}{An mle_problem object}
}
\value{
Fisher information function
}
\description{
Returns the Fisher information matrix function, computing numerically if not provided.
}
