% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem.R
\name{get_score}
\alias{get_score}
\title{Get score function from problem}
\usage{
get_score(problem)
}
\arguments{
\item{problem}{An mle_problem object}
}
\value{
Score function
}
\description{
Returns the score (gradient) function, computing numerically if not provided.
}
