% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{is_mle_config}
\alias{is_mle_config}
\title{Check if object is an mle_config}
\usage{
is_mle_config(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical
}
\description{
Check if object is an mle_config
}
