% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{is_mle_constraint}
\alias{is_mle_constraint}
\title{Check if object is an mle_constraint}
\usage{
is_mle_constraint(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical
}
\description{
Check if object is an mle_constraint
}
