% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{is_mle_numerical}
\alias{is_mle_numerical}
\title{is_mle_numerical}
\usage{
is_mle_numerical(x)
}
\arguments{
\item{x}{the `mle` object}

\item{...}{additional arguments to pass}
}
\description{
Function to determine whether an object `x` is of type `mle_numerical`.
}
