% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem.R
\name{is_mle_problem}
\alias{is_mle_problem}
\title{Check if object is an mle_problem}
\usage{
is_mle_problem(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical
}
\description{
Check if object is an mle_problem
}
