% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{is_tracing}
\alias{is_tracing}
\title{Check if tracing is enabled}
\usage{
is_tracing(trace)
}
\arguments{
\item{trace}{An mle_trace object}
}
\value{
Logical indicating if any tracing is enabled
}
\description{
Check if tracing is enabled
}
