% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_local_search.R
\name{mle_local_search}
\alias{mle_local_search}
\title{mle_local_search}
\usage{
mle_local_search(dir, theta0, loglike = NULL, options = list())
}
\arguments{
\item{dir}{function, promising direction function}

\item{theta0}{numeric, initial guess}

\item{options}{list, options for the local search, see function description.}
}
\value{
an `mle` object with additional attributes `iter` and `converged`
        and optionally `path` if `trace` is TRUE.
}
\description{
Performs a local search to find the MLE, assuming the MLE is an
interior point of the support and that an initial guess `theta0`
that is near the MLE is provided. Use a global search method
like `sim_anneal` to find a good initial guess.
}
\section{Functions}{
\itemize{
\item \code{mle_local_search()}: options

}}
\section{Fields}{

\describe{
\item{\code{sup}}{function, domain of support for log-likelihood}

\item{\code{eta}}{numeric, learning rate, defaults to 1}

\item{\code{max_iter}}{integer, maximum number of iterations, defaults to 1000}

\item{\code{max_iter_ls}}{integer, maximum number of iterations for the
line search, defaults to 1000}

\item{\code{abs_tol}}{numeric, tolerance for convergence, defaults to NULL
(use rel_tol instead)}

\item{\code{rel_tol}}{numeric, relative tolerance for convergence, defaults to 1e-5}

\item{\code{r}}{numeric, backtracking line search parameter, defaults to 0.5}

\item{\code{proj}}{function, projection function to enforce domain of support}

\item{\code{norm}}{function, distance measure for convergence checks, defaults
to the to the infinity norm.}

\item{\code{debug}}{logical, output debugging information if TRUE; default FALSE}

\item{\code{trace}}{logical, if TRUE store the path of the search in the `path`
attribute of the output; default FALSE}

\item{\code{line_search}}{logical, if TRUE, perform a line search; default TRUE
in this case, learning rate `eta` refers to the maximum step size
that can be taken per iteration.}

\item{\code{debug_freq}}{integer, frequency of debug output, defaults to 1}
}}

