% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_newton_raphson.R
\name{mle_newton_raphson}
\alias{mle_newton_raphson}
\title{mle_newton_raphson}
\usage{
mle_newton_raphson(score, fim, theta0, inverted = FALSE, options = list())
}
\arguments{
\item{score}{score Function, the gradient of log-likelihood}

\item{fim}{FIM function}

\item{theta0}{initial guess of MLE}

\item{inverted}{logical, if TRUE `fim` is covariance instead of FIM}
}
\value{
an object of class `mle_newton_raphson`, which is an `mle` object
}
\description{
Find an MLE using the Newton-Raphson method.
}
