% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{mle_numerical}
\alias{mle_numerical}
\title{mle_numerical}
\usage{
mle_numerical(theta.hat, loglike, score, info, sigma, iter, converged)
}
\description{
a constructor for the mle_numerical class.
}
