% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_random_restart.R
\name{mle_random_restart}
\alias{mle_random_restart}
\title{mle_random_restart}
\usage{
mle_random_restart(rtheta0, options)
}
\arguments{
\item{rtheta0}{initial guess of MLE distribution, defaults to MVN with wide variance}
}
\description{
MLE method using random restart
}
\details{
An interesting idea is to use random restarts to find the MLE. This is
similar to the idea of simulated annealing, but instead of using a
temperature parameter, we just try a bunch of random starting and then try
to find the MLE from each of those starting points using `mle_solver`. The
best MLE is returned.

A reasonable `rtheta0` is a multivariate normal distribution with a wide
variance. This is because we want to explore the entire parameter space.

If we already have a candidate MLE `x`, we can use that as the sampling distribution of
the MLE (MVN with mean `point(x)` and variance-covariance `vcov(x)`). This
will generate random starting points that are close to the candidate MLE, which
may help us find a better MLE.
}
