% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_sim_anneal.R
\name{mle_sim_anneal}
\alias{mle_sim_anneal}
\title{mle_sim_anneal}
\usage{
mle_sim_anneal(theta0, loglike = NULL, options = list())
}
\description{
This function takes the output of `sim_anneal` and turns it into an `mle` object.
}
