% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_solve.R
\name{mle_solve}
\alias{mle_solve}
\title{mle_solve}
\usage{
mle_solve(options, theta0, method = c("optim"), ...)
}
\arguments{
\item{theta0}{numeric vector, initial guess for MLE}
}
\description{
Uses various functions and algorithms to find the MLE.
}
