% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{new_trace_recorder}
\alias{new_trace_recorder}
\title{Create a trace recorder}
\usage{
new_trace_recorder(trace, n_params)
}
\arguments{
\item{trace}{mle_trace configuration}

\item{n_params}{Number of parameters (for pre-allocation)}
}
\value{
A trace recorder environment
}
\description{
Internal function to create a mutable trace recorder.
}
\keyword{internal}
