% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{normal_sampler}
\alias{normal_sampler}
\title{Normal Sampler Factory}
\usage{
normal_sampler(center, sd = 1)
}
\arguments{
\item{center}{Mean of the normal distribution}

\item{sd}{Standard deviation (scalar or vector)}
}
\value{
A sampler function
}
\description{
Creates a sampler function for use with \code{with_restarts} that
generates normally distributed starting points around a center.
}
\examples{
sampler <- normal_sampler(c(0, 1), sd = c(5, 0.5))
strategy <- gradient_ascent() \%>\% with_restarts(n = 20, sampler = sampler)

}
