% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical.mle.R
\docType{package}
\name{numerical.mle}
\alias{numerical.mle}
\title{`numerical.mle`: A package for numerically solving maximum likelihood estimators
from log-likelihood functions.}
\description{
The object representing a fitted model is a type of `mle` object, the maximum
likelihood estimator of the model with respect to observed data. We use the
R package `algebraic.mle` to represent these objects.
}
