% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{record_iteration}
\alias{record_iteration}
\title{Record an iteration to trace}
\usage{
record_iteration(recorder, theta, value = NULL, gradient = NULL)
}
\arguments{
\item{recorder}{Trace recorder from new_trace_recorder}

\item{theta}{Current parameters}

\item{value}{Current log-likelihood (or NULL)}

\item{gradient}{Current gradient (or NULL)}
}
\description{
Record an iteration to trace
}
\keyword{internal}
