% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{stochastic_loglike}
\alias{stochastic_loglike}
\title{stochastic loglikelihood constructor
good for large datasets. if applied to a gradient ascent method, this
will perform stochastic gradient ascent.}
\usage{
stochastic_loglike(log.p, obs, options)
}
\arguments{
\item{log.p}{log pdf (or pmf) of the parametric model being fit to `obs`
parameters. it can also just be proportional to the log pdf, since sometimes
the normalizing constant is unknown or hard to compute.}

\item{obs}{a matrix, vector, or data frame of observations}

\item{options}{a list of options}
}
\description{
stochastic loglikelihood constructor
good for large datasets. if applied to a gradient ascent method, this
will perform stochastic gradient ascent.
}
