% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_grid_search.R
\name{subdivide_region}
\alias{subdivide_region}
\title{subdivide_region}
\usage{
subdivide_region(lower, upper, grid_size)
}
\arguments{
\item{lower}{lower bounds of parameter support (vector)}

\item{upper}{upper bounds of parameter support (vector)}

\item{grid_size}{maximum size of each dimension of each hypercube
that makes up the grid}
}
\description{
subdivide_region
}
