% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{uniform_sampler}
\alias{uniform_sampler}
\title{Uniform Sampler Factory}
\usage{
uniform_sampler(lower, upper)
}
\arguments{
\item{lower}{Lower bounds for each parameter}

\item{upper}{Upper bounds for each parameter}
}
\value{
A sampler function
}
\description{
Creates a sampler function for use with \code{with_restarts} that
generates uniformly distributed starting points.
}
\examples{
sampler <- uniform_sampler(c(-10, 0.1), c(10, 5))
strategy <- gradient_ascent() \%>\% with_restarts(n = 20, sampler = sampler)

}
