% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem.R
\name{update.mle_problem}
\alias{update.mle_problem}
\title{Update an mle_problem}
\usage{
\method{update}{mle_problem}(object, ...)
}
\arguments{
\item{object}{An mle_problem}

\item{...}{Named arguments to update}
}
\value{
New mle_problem
}
\description{
Create a new problem with some fields updated.
}
