% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{is_mle_numerical}
\alias{is_mle_numerical}
\title{Check if object is an mle_numerical}
\usage{
is_mle_numerical(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical
}
\description{
Check if object is an mle_numerical
}
