% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_gradient_ascent.R
\name{mle_gradient_ascent}
\alias{mle_gradient_ascent}
\title{mle_gradient_ascent}
\usage{
mle_gradient_ascent(theta0, score, options)
}
\arguments{
\item{theta0}{initial guess of MLE}

\item{score}{score function, the gradient of a log-likelihood}
}
\value{
an object of class `mle_gradient_ascent`, which is an `mle` object
}
\description{
MLE method using gradient ascent
}
