% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{num_iterations}
\alias{num_iterations}
\title{Get number of iterations}
\usage{
num_iterations(x, ...)
}
\arguments{
\item{x}{An mle result object}

\item{...}{Additional arguments (unused)}
}
\value{
Number of iterations
}
\description{
Get number of iterations
}
